workers/builder/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
pydantic==2.10.5
redis==5.2.1
psycopg2-binary==2.9.11
pyelftools==0.31
